==> robustness_check/Part3_Nonp_merge.Rmd <==
---
Project: "Controlling for the general psychopathology factor p in genomic research on psychiatric disorders"
Title of the script: "Part3_Nonp_merge"
author: "Wangjingyi Liao & Engin Keser"
date: "2023-11-22"
output: html_document
---
Project: "Controlling for the general psychopathology factor p in genomic research on psychiatric disorders"

Title of the script: "Part3_Nonp_merge"

```{r setup, include=FALSE}
knitr::opts_chunk$set(eval = FALSE)
```
p gwas
```{r}

setwd('/scratch/prj/teds/Sumstate_PRS_Codelab/disease_specificity/robustness_check/p_gwas')

getwd()

library(dplyr)
library(qqman)
library(data.table)

# get list of all .rds files in directory
#files <- list.files(pattern = "\\.rds$")

files <- list.files(pattern = "^p_gwas")

load(files[1])

test <- result

colnames <- colnames(test)

remove(test)

# initialize an empty data frame
combined_data <- data.frame(matrix(ncol = 27, nrow = 0))

colnames(combined_data) <- colnames

# loop through each file, read in data, and bind to existing data frame
for (file in files) {
  load(file)
  combined_data <- rbind(combined_data, result)
}

saveRDS(combined_data, file = 'p_gwas_results.rds')

head(combined_data)

colnames(combined_data)[colnames(combined_data) == "Pval_Estimate"] <- "P"

num_rows <- nrow(combined_data)
#restrict to MAF of 40% and 10%
combined_data2<-subset(combined_data, combined_data$MAF <= .4 & combined_data$MAF >= .1)
#calculate expected sample size (N_hat)
N_hat<-mean(1/((2*combined_data2$MAF*(1-combined_data2$MAF))*combined_data2$se_c^2))
print(paste("Number of rows in p gwas", ":", num_rows, N_hat))

fwrite(combined_data, file = "p_gwas_results.txt", sep = "\t", col.names = TRUE)
#manhattan(combined_data, p = 'P')

```

```{r}

setwd('/scratch/prj/teds/Sumstate_PRS_Codelab/robustness_check/')

getwd()

library(dplyr)
library(qqman)
library(data.table)

# get list of all .rds files in directory
#files <- list.files(pattern = "\\.rds$")

files <- list.files(pattern = "^SCZnonp")

load(files[1])

test <- result[[1]]

colnames <- colnames(test)

remove(test)

# initialize an empty data frame
combined_data <- data.frame(matrix(ncol = 27, nrow = 0))

colnames(combined_data) <- colnames

# loop through each file, read in data, and bind to existing data frame
for (file in files) {
  load(file)
  test <- result[[1]]
  combined_data <- rbind(combined_data, test)
}

#save as RDS
saveRDS(combined_data, file = 'SCZ_nonp_gwas_results.rds')

#basic manhattan visualization, better manhattan plot script is in script Part 7
manhattan(combined_data, p = 'Pval_Estimate')


```
Bipolar Nonp
```{r}

# get list of all .rds files in directory
#files <- list.files(pattern = "\\.rds$")

files <- list.files(pattern = "^BIPnonp")

load(files[1])

test <- result[[1]]

colnames <- colnames(test)

remove(test)

# initialize an empty data frame
combined_data <- data.frame(matrix(ncol = 27, nrow = 0))

colnames(combined_data) <- colnames

# loop through each file, read in data, and bind to existing data frame
for (file in files) {
  load(file)
  test <- result[[1]]
  combined_data <- rbind(combined_data, test)
}

saveRDS(combined_data, file = 'BIP_nonp_gwas_results.rds')



manhattan(combined_data, p = 'Pval_Estimate')


```
ADHD Nonp
```{r}

# get list of all .rds files in directory
#files <- list.files(pattern = "\\.rds$")

files <- list.files(pattern = "^ADHDnonp")

load(files[1])

test <- result[[1]]

colnames <- colnames(test)

remove(test)

# initialize an empty data frame
combined_data <- data.frame(matrix(ncol = 27, nrow = 0))

colnames(combined_data) <- colnames

# loop through each file, read in data, and bind to existing data frame
for (file in files) {
  load(file)
  test <- result[[1]]
  combined_data <- rbind(combined_data, test)
}

saveRDS(combined_data, file = 'ADHD_nonp_gwas_results.rds')

manhattan(combined_data, p = 'Pval_Estimate')


```
PTSD Nonp
```{r}

# get list of all .rds files in directory
#files <- list.files(pattern = "\\.rds$")

files <- list.files(pattern = "^PTSDnonp")

load(files[1])

test <- result[[1]]

colnames <- colnames(test)

remove(test)

# initialize an empty data frame
combined_data <- data.frame(matrix(ncol = 27, nrow = 0))

colnames(combined_data) <- colnames

# loop through each file, read in data, and bind to existing data frame
for (file in files) {
  load(file)
  test <- result[[1]]
  combined_data <- rbind(combined_data, test)
}

saveRDS(combined_data, file = 'PTSD_nonp_gwas_results.rds')

manhattan(combined_data, p = 'Pval_Estimate')


```
MDD Nonp
```{r}

# get list of all .rds files in directory
#files <- list.files(pattern = "\\.rds$")

files <- list.files(pattern = "^MDDnonp")

load(files[1])

test <- result[[1]]

colnames <- colnames(test)

remove(test)

# initialize an empty data frame
combined_data <- data.frame(matrix(ncol = 27, nrow = 0))

colnames(combined_data) <- colnames

# loop through each file, read in data, and bind to existing data frame
for (file in files) {
  load(file)
  test <- result[[1]]
  combined_data <- rbind(combined_data, test)
}

saveRDS(combined_data, file = 'MDD_nonp_gwas_results.rds')

manhattan(combined_data, p = 'Pval_Estimate')


```
ANX Nonp
```{r}

# get list of all .rds files in directory
#files <- list.files(pattern = "\\.rds$")

files <- list.files(pattern = "^ANXnonp")

load(files[1])

test <- result[[1]]

colnames <- colnames(test)

remove(test)

# initialize an empty data frame
combined_data <- data.frame(matrix(ncol = 27, nrow = 0))

colnames(combined_data) <- colnames

# loop through each file, read in data, and bind to existing data frame
for (file in files) {
  load(file)
  test <- result[[1]]
  combined_data <- rbind(combined_data, test)
}

saveRDS(combined_data, file = 'ANX_nonp_gwas_results.rds')

manhattan(combined_data, p = 'Pval_Estimate')


```
ASD Nonp
```{r}

# get list of all .rds files in directory
#files <- list.files(pattern = "\\.rds$")

files <- list.files(pattern = "^ASDnonp")

load(files[1])

test <- result[[1]]

colnames <- colnames(test)

remove(test)

# initialize an empty data frame
combined_data <- data.frame(matrix(ncol = 27, nrow = 0))

colnames(combined_data) <- colnames

# loop through each file, read in data, and bind to existing data frame
for (file in files) {
  load(file)
  test <- result[[1]]
  combined_data <- rbind(combined_data, test)
}

saveRDS(combined_data, file = 'ASD_nonp_gwas_results.rds')

manhattan(combined_data, p = 'Pval_Estimate')


```

OCD Nonp
```{r}

# get list of all .rds files in directory
#files <- list.files(pattern = "\\.rds$")

files <- list.files(pattern = "^OCDnonp")

load(files[1])

test <- result[[1]]

colnames <- colnames(test)

remove(test)

# initialize an empty data frame
combined_data <- data.frame(matrix(ncol = 27, nrow = 0))

colnames(combined_data) <- colnames

# loop through each file, read in data, and bind to existing data frame
for (file in files) {
  load(file)
  test <- result[[1]]
  combined_data <- rbind(combined_data, test)
}

saveRDS(combined_data, file = 'OCD_nonp_gwas_results.rds')

manhattan(combined_data, p = 'Pval_Estimate')


```
ALCH Nonp
```{r}

# get list of all .rds files in directory
#files <- list.files(pattern = "\\.rds$")

files <- list.files(pattern = "^ALCHnonp")

load(files[1])

test <- result[[1]]

colnames <- colnames(test)

remove(test)

# initialize an empty data frame
combined_data <- data.frame(matrix(ncol = 27, nrow = 0))

colnames(combined_data) <- colnames

# loop through each file, read in data, and bind to existing data frame
for (file in files) {
  load(file)
  test <- result[[1]]
  combined_data <- rbind(combined_data, test)
}

saveRDS(combined_data, file = 'ALCH_nonp_gwas_results.rds')

manhattan(combined_data, p = 'Pval_Estimate')


```
AN Nonp
```{r}

# get list of all .rds files in directory
#files <- list.files(pattern = "\\.rds$")

files <- list.files(pattern = "^ANnonp")

load(files[1])

test <- result[[1]]

colnames <- colnames(test)

remove(test)

# initialize an empty data frame
combined_data <- data.frame(matrix(ncol = 27, nrow = 0))

colnames(combined_data) <- colnames

# loop through each file, read in data, and bind to existing data frame
for (file in files) {
  load(file)
  test <- result[[1]]
  combined_data <- rbind(combined_data, test)
}

saveRDS(combined_data, file = 'AN_nonp_gwas_results.rds')

manhattan(combined_data, p = 'Pval_Estimate')


```

TS Nonp
```{r}

# get list of all .rds files in directory
#files <- list.files(pattern = "\\.rds$")

files <- list.files(pattern = "^TSnonp")

load(files[1])

test <- result[[1]]

colnames <- colnames(test)

remove(test)

# initialize an empty data frame
combined_data <- data.frame(matrix(ncol = 27, nrow = 0))

colnames(combined_data) <- colnames

# loop through each file, read in data, and bind to existing data frame
for (file in files) {
  load(file)
  test <- result[[1]]
  combined_data <- rbind(combined_data, test)
}

saveRDS(combined_data, file = 'TS_nonp_gwas_results.rds')

manhattan(combined_data, p = 'Pval_Estimate')


```

in bash, copy all rds files to the nonp_residual_sumstats folder
```{bash}
#mkdir nonp_residual_sumstats
cp *_nonp_gwas_results.rds ./nonp_residuals_sumstats

```

all rds file convert to txt file and check the SNP number to check if missing batches. change the column names and calcualte the expected sample size of Nonps using Mallard 2019 method

```{r}
setwd('./nonp_residuals_sumstats/')

library(data.table)
library(tidyr)

traits <- c('SCZ', 'BIP' , 'ADHD', 'MDD', 'ANX', 'ASD', 'PTSD', 'ALCH', 'OCD', 'AN', 'TS')

# Create an empty list to store the summary results
Neff_list <- list()

# Iterate over each trait
for (i in traits) {
  trait <- readRDS(paste0(i, "_nonp_gwas_results.rds"))
  head(trait)
  colnames(trait)[colnames(trait) == "Pval_Estimate"] <- "P"
  fwrite(trait, file = paste0(i, "_nonp_gwas_results.txt"), sep = "\t", col.names = TRUE)
  num_rows <- nrow(trait)
  #restrict to MAF of 40% and 10%
  trait2<-subset(trait, trait$MAF <= .4 & trait$MAF >= .1)
  #calculate expected sample size (N_hat)
  N_hat<-mean(1/((2*trait2$MAF*(1-trait2$MAF))*trait2$SE^2))
  print(paste("Number of rows in", i, ":", num_rows, N_hat))
}

```
